ref(chat): Enforce service Slack boundary
Move Slack-backed conversation image and backfill details out of domain services so services depend on small injected ports instead of Slack infrastructure. Add an architecture guard and update the contract docs to keep the boundary enforceable.
