+++ specs/chat-architecture-spec.md
@@ -3,7 +3,7 @@
 ## Metadata
 
 - Created: 2026-03-21
-- Last Edited: 2026-03-22
+- Last Edited: 2026-05-09
 
 ## Changelog
 
@@ -14,6 +14,7 @@
 - 2026-03-22: Moved canonical ingress routing and chat bindings under `chat/ingress/*` and deleted the old `chat-background-patch.ts` module.
 - 2026-03-22: Removed test-only plugin registry mutation APIs and made plugin discovery reload from the current root signature instead of import-time state.
 - 2026-03-22: Replaced prototype-patch ingress with the explicit `JuniorChat` subclass, removed the OAuth resume post-message observer global, split queue transport from queue retry policy, and replaced the ambient user-token-store singleton with construction at the call site.
+- 2026-05-09: Added the enforced service-to-Slack boundary: domain services depend on injected Slack-backed ports instead of Slack infrastructure modules.
 
 ## Status
 
@@ -64,6 +65,7 @@ Define the normative architecture contract for `packages/junior/src/chat` so new
 - Non-`app/` modules must not import from `app/`.
 - `runtime/` may depend on `services/`, `state/`, `queue/`, `turn/`, `slack/`, and pure helpers.
 - `services/` must not depend on `runtime/`.
+- `services/` must not import Slack infrastructure; use small injected ports owned by the service when a domain service needs Slack-backed data or files.
 - `state/` must not depend on `runtime/` or service modules.
 - `ingress/` may route into queue/runtime entrypoints, but must not own business logic that belongs in `runtime/` or `services/`.
